+++ tests/Units/TagItemTest.php
@@ -31,6 +31,7 @@
 namespace GlpiPlugin\Tag\Tests\Units;
 
 use GlpiPlugin\Tag\Tests\TagTestCase;
+use PluginTagTag;
 use Ticket;
 
 final class TagItemTest extends TagTestCase
@@ -56,4 +57,36 @@ final class TagItemTest extends TagTestCase
         $this->isItemTagged($ticket, $tagID2);
     }
 
+    public function testTagOutOfEntityScopeIsNotLinked(): void
+    {
+        $this->login();
+
+        $out_of_scope_entity = getItemByTypeName('Entity', '_test_child_2', true);
+        $tag = new PluginTagTag();
+        $tag->add([
+            'name' => 'OutOfScopeTag',
+            'is_active' => 1,
+            'type_menu' => ['Ticket'],
+            'entities_id' => $out_of_scope_entity,
+            'is_recursive' => 0,
+        ]);
+        $tagID = $tag->getID();
+        $this->assertGreaterThan(0, $tagID);
+
+        $this->setEntity('_test_child_1', false);
+
+        $ticket = new Ticket();
+        $ticket->add([
+            'name' => 'Ticket out of scope tag',
+            'content' => 'Ticket out of scope tag',
+            'entities_id' => getItemByTypeName('Entity', '_test_child_1', true),
+            '_plugin_tag_tag_process_form' => 1,
+            '_plugin_tag_tag_values'   => [
+                $tagID,
+            ],
+        ]);
+        $this->assertGreaterThan(0, $ticket->getID());
+
+        $this->isItemNotTagged($ticket, $tagID);
+    }
 }